+++ railties/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Add `rails test:mailboxes`.
+
+    *George Claghorn*
+
 *   Introduce guard against DNS rebinding attacks
 
     The `ActionDispatch::HostAuthorization` is a new middleware that prevent